refactor: consolidate get_mcp_server() singleton into web/api/__init__
Four identical copies across profiles, ensembles, scripts, and
artifacts modules replaced with a single shared implementation.
