+++ src/llm_orc/web/api/artifacts.py
@@ -7,19 +7,9 @@ from typing import Any
 
 from fastapi import APIRouter, HTTPException
 
-from llm_orc.mcp import MCPServer
+from llm_orc.web.api import get_mcp_server
 
 router = APIRouter(prefix="/api/artifacts", tags=["artifacts"])
-
-_mcp_server: MCPServer | None = None
-
-
-def get_mcp_server() -> MCPServer:
-    """Get or create the MCP server instance."""
-    global _mcp_server
-    if _mcp_server is None:
-        _mcp_server = MCPServer()
-    return _mcp_server
 
 
 @router.get("")
